% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordup-package.R
\docType{package}
\name{wordup-package}
\alias{wordup}
\alias{wordup-package}
\title{wordup: Convert Word to Govspeak}
\description{
Try to convert a Word document (docx) to the equivalent Govspeak Markdown, ready for upload to the UK government's publishing platform.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://matt-dray.github.io/wordup/}
  \item Report bugs at \url{https://github.com/matt-dray/wordup/issues}
}

}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray@gmail.com}

}
\keyword{internal}
