% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{wu_body}
\alias{wu_body}
\title{Extract Specific Body Elements}
\usage{
wu_body(doc_list, element = c("p", "tbl"))
}
\arguments{
\item{doc_list}{List. Output from \link{wu_read}.}

\item{element}{Character. The elements you want to return.}
}
\value{
A list with an element for each instance of the desired element.
}
\description{
Extract Specific Body Elements
}
\examples{
path <- system.file("examples/simple.docx", package = "wordup")
doc_list <- wu_read(path)
p_list <- wu_body(doc_list, "p")
str(p_list, give.attr = FALSE, max.level = 1)

}
