% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{wu_p}
\alias{wu_p}
\title{Extract All 'p' Body Text and Style to a Dataframe}
\usage{
wu_p(p_list)
}
\arguments{
\item{p_list}{List. Output from \link{wu_body} with argument \code{element = "p"}.}
}
\value{
A data.frame with a row per 'p' element and columns with text and
possibly style information.
}
\description{
Extract All 'p' Body Text and Style to a Dataframe
}
\examples{
path <- system.file("examples/simple.docx", package = "wordup")
doc_list <- wu_read(path)
p_list <- wu_body(doc_list, "p")
wu_p(p_list)

}
