% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{wu_read}
\alias{wu_read}
\title{Read a Word File to a List}
\usage{
wu_read(docx_path)
}
\arguments{
\item{docx_path}{Character. A path to a docx file.}
}
\value{
A nested list.
}
\description{
Unzips a docx file, reads the XML from \verb{/word/document.xml} and converts it
to a list object for further processing.
}
\examples{
path <- system.file("examples/simple.docx", package = "wordup")
body_list <- wu_read(path)
str(body_list, give.attr = FALSE, max.level = 3)

}
