% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{convert_table_to_md}
\alias{convert_table_to_md}
\title{Convert a Copy-Pasted Word Table to Govspeak}
\usage{
convert_table_to_md(
  pasted_table = clipr::read_clip_tbl(),
  guess_types = TRUE,
  ignore_regex = ",|\%|\\\\[c\\\\]",
  has_row_titles = FALSE,
  totals_rows = NULL,
  to_clipboard = TRUE
)
}
\arguments{
\item{pasted_table}{Character. A single string that contains the content of
a table copy and pasted from a Word document.}

\item{guess_types}{Logical. Should data types be guessed for each column
based on their content? Defaults to \code{TRUE}. If \code{FALSE}, all columns will
be returned as character type.}

\item{ignore_regex}{Character. A regular expression of strings to ignore
when trying to guess column types.}

\item{has_row_titles}{Logical. Should the first column be treated as though
it contains titles for each row? Defaults to \code{FALSE}. If \code{TRUE}, the
first column will be marked-up as bold.}

\item{totals_rows}{Integer. A vector of indices to identify rows that
contain totals. These will marked up as bold.}

\item{to_clipboard}{Logical. Should the output be copied to your clipboard?
Defaults to \code{TRUE}.}
}
\value{
Character. A string that contains Govspeak Markdown that represents
the copy-pasted table.
}
\description{
Provide a copied table from a Word document and be returned a Govspeak
Markdown version of it. Some post-editing may be necessary for more complex
tables.
}
\examples{
pasted_table <- c("Column 1	Column 2	Column 3	Column 4	Column 5
X	100	1,000	1\%	15
Y	200	2,000	2\%	12
Z	300	3,000	3\%	[c]")

convert_table_to_md(pasted_table, to_clipboard = FALSE)

}
