% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{wu_p}
\alias{wu_p}
\title{Extract All 'p' Body Text and Style to a Dataframe}
\usage{
wu_p(body_list)
}
\arguments{
\item{body_list}{List. Output from \link{wu_body}.}
}
\value{
A data.frame with a row per 'p' element and columns with text and
possibly style information.
}
\description{
Extract All 'p' Body Text and Style to a Dataframe
}
\examples{
path <- system.file("examples/simple.docx", package = "wordup")
body_list <- wu_read(path)
p_list <- wu_body(body_list, "p")
head(wu_p(p_list))

}
